Introduce explicit build mode and clean up code
This commit introduces an explicit build mode in get-mods.sh that determines
how the versions of external modules are selected.

Possible modes are:
 - release: chooses latest release for each module
 - snapshot: uses a snapshot/development version for each module (if available)

In action.bash, the mode is automatically set based on the GITHUB_REF variable.
For tags starting with "v" (i.e. version tags like v2.9.0) it is release, other
tag names are not supported. For branches it is snapshot.